feat/Client : refacto as an accountant update categories
